models modification

+++ main.py
@@ -49,18 +49,23 @@ async def read_user(user_id: int, db: Session = Depends(get_db)):
 
 # Update Users Endpoint
 @app.put("/users/{user_id}", response_model=schemas.User)
-def update_user(user_id: int, user_update: schemas.UserUpdate, db: Session = Depends(get_db)):
+def update_user(user_id: int, user_update: schemas.UserUpdate, db: Session = Depends(get_db), current_user: dict = Depends(auth.get_current_user)):
     db_user = crud.get_user(db=db, user_id=user_id)
     if db_user is None:
         raise HTTPException(status_code=404, detail="Users not found")
+    if db_user.id != current_user["id"]:
+        raise HTTPException(status_code=403, detail="You are not allowed to update this user")
     updated_user = crud.update_user(db=db, user=db_user, user_update=user_update)
     return updated_user
 
 
 # Delete Users Endpoint
 @app.delete("/users/{user_id}")
-def delete_user(user_id: int, db: Session = Depends(get_db)):
-    deleted = crud.delete_user(db=db, user_id=user_id)
-    if not deleted:
-        raise HTTPException(status_code=404, detail="Users not found")
-    return {"message": "Users deleted successfully"}
+def delete_user(user_id: int, db: Session = Depends(get_db), current_user: dict = Depends(auth.get_current_user)):
+    if user_id != current_user["id"]:
+        raise HTTPException(status_code=403, detail="You are not allowed to delete this user")
+    else:
+        deleted = crud.delete_user(db=db, user_id=user_id)
+        if not deleted:
+            raise HTTPException(status_code=404, detail="Users not found")
+        return {"message": "Users deleted successfully"}
